Adds heartbeat to keep websocket alive
By default, the phoenix server will close a websocket from which it has
not received data.  When this happens, elm websocket reconnects.  To the
phoenix server, this is a new socket.  Any state associated with the
prior socket (for example, channels that were joined) are lost.

phoenix.js that comes bundled with phoenix uses an identical heartbeat
mechanism to keep the websocket alive.

You can see that here:
https://github.com/phoenixframework/phoenix/blob/master/web/static/js/phoenix.js